Added TimeGenerated and timestamp columns
I updated a project statement to preserve the TimeGenerated column.  I did this so that interesting events can be bookmarked and tracked by time.  If we think this will add too much noise, due to adding non-distinct entries, please let me know, and block the change.

Added timestamp mapping (| extend timestamp = TimeGenerated) so that bookmarks will track the event time in the bookmarked row by default vs. the time the bookmark was created.